+++ public/js/layers.js
@@ -22,6 +22,7 @@
     correlationHeatmap: true,
     cvd:                true,
     optionsLevels:      true,
+    ddBands:            true,
   };
 
   function load() {